+++ Tensile/Source/client/main.cpp
@@ -492,26 +492,6 @@ namespace Tensile
     } // namespace Client
 } // namespace Tensile
 
-inline void print_memory_size(size_t memory_size)
-{
-    if(memory_size < 1024)
-    {
-        std::cout << std::setprecision(0) << memory_size << " Bytes";
-    }
-    else if(memory_size < 1048576)
-    {
-        std::cout << std::setprecision(3) << float(memory_size) / 1024.0f << " KB";
-    }
-    else if(memory_size < 1073741824)
-    {
-        std::cout << std::setprecision(6) << float(memory_size) / 1048576.0f << " MB";
-    }
-    else
-    {
-        std::cout << std::setprecision(9) << float(memory_size) / 1073741824.0f << " GB";
-    }
-}
-
 size_t calculate_flush_count(size_t                                       arg_flush_count,
                              size_t                                       arg_flush_memory_size,
                              Tensile::Client::ClientProblemFactory const& problemFactory)
@@ -542,7 +522,7 @@ size_t calculate_flush_count(size_t                                       arg_fl
         flush_count = arg_flush_count;
 
         std::cout << "flush_memory_size = ";
-        print_memory_size(flush_count * cached_size);
+        Tensile::print_memory_size(flush_count * cached_size);
         std::cout << std::endl;
     }
     else if(arg_flush_memory_size != default_arg_flush_memory_size)
